move ROI-related functionality to its own module
